feat(seal): refactor seal calls
