+++ crates/executor/src/derive.rs
@@ -9,3 +9,6 @@ pub use crate::{
 
 /// Host function parcel
 pub type HostParcel<M, F, T> = (M, F, HostFuncType<T>);
+
+/// Custom SealCall
+pub type SealCall<T> = HostParcel<&'static str, &'static str, T>;
